% Generated by build script
\name{cran_deps}
\alias{cran_deps}
\title{Get reverse dependencies for a CRAN package}
\usage{
cran_deps()
}
\arguments{
\item{package}{Package name}
}
\value{tibble: package, type (Depends/Imports/Suggests/LinkingTo), version, title}
\description{Get reverse dependencies for a CRAN package}
