% Generated by build script
\name{cran_search}
\alias{cran_search}
\title{Search CRAN packages}
\usage{
cran_search()
}
\arguments{
\item{query}{Search term}
\item{size}{Number of results (default 30)}
}
\value{tibble: package, version, title, description, maintainer}
\description{Search CRAN packages}
