% Generated by build script
\name{cran_top}
\alias{cran_top}
\title{Get top downloaded CRAN packages}
\usage{
cran_top()
}
\arguments{
\item{count}{Number of top packages to return (default 25)}
\item{period}{"last-day", "last-week", or "last-month" (default)}
}
\value{tibble: rank, package, downloads}
\description{Get top downloaded CRAN packages}
